Add FFI functions to get and read a wallets Seed Words
This PR adds the FFI functions to get the seed words that represent the seed private key of the key manager that wallet is using. This PR also provides the FFI functions used to restructure the collection of seed words and free the memory when you are done with it.